EmailApp/requirements.txt
python-dotenv
yagmail